M2 - moving models to module
