+++ musik_lib/admin.py
@@ -3,7 +3,7 @@ from django.contrib import admin
 # Register your models here.
 from django.contrib import admin
 
-from musik_lib.models import *
+from musik_lib.models.base import *
 
 admin.site.register(Artist)
 admin.site.register(Track)
