+++ musik_lib/models/base.py
@@ -51,7 +51,10 @@ class Collection(models.Model):
         )
 
     def number_of_tracks(self):
-        return len(self.track_set.all())
+        if hasattr(self, "track_set"):
+            return len(self.track_set.all())
+        else:
+            return 0
 
 
 class Track(models.Model):
@@ -79,67 +82,3 @@ class Artist(models.Model):
 
     def __str__(self):
         return self.name
-
-
-class DuplicateTrack(models.Model):
-    track = models.OneToOneField(
-        Track,
-        on_delete=models.CASCADE,
-        primary_key=True,
-    )
-
-
-class LibraryStat(models.Model):
-    library = models.OneToOneField(
-        Library,
-        on_delete=models.CASCADE,
-        primary_key=True,
-    )
-
-
-class CollectionStat(models.Model):
-    collection = models.OneToOneField(
-        Collection,
-        on_delete=models.CASCADE,
-        primary_key=True,
-    )
-    library_stat = models.ForeignKey(
-        LibraryStat,
-        on_delete=models.CASCADE,
-        default=None,
-    )
-
-    duplicate_tracks = models.ManyToManyField(DuplicateTrack)
-
-
-class ArtistFrequencyCollection(models.Model):
-    artist = models.ForeignKey(
-        Artist,
-        on_delete=models.CASCADE,
-    )
-    collection_stat = models.ForeignKey(
-        CollectionStat,
-        on_delete=models.CASCADE
-    )
-    frequency = models.PositiveSmallIntegerField(default=1)
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=['artist', 'collection_stat'], name='Artist And Collection')
-        ]
-
-    def __str__(self):
-        return "{} - {} - {}".format(self.artist.name, self.collection_stat.collection.name, self.frequency)
-
-
-class ArtistFrequencyLibrary(models.Model):
-    artist = models.ForeignKey(
-        Artist,
-        on_delete=models.CASCADE,
-    )
-    library_stat = models.ForeignKey(
-        LibraryStat,
-        on_delete=models.CASCADE
-    )
-    frequency = models.PositiveSmallIntegerField(default=1)
